Tweak docs, separate page/url info from page scan info
