+++ packages/storage-documents/src/scan-types.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-export type ScanType = 'a11y' | 'cookies';
+export type ScanType = 'a11y' | 'privacy' | 'security';
 
 export type ScanStatus = 'pass' | 'fail' | 'error' | 'pending';
