+++ packages/storage-documents/src/website-scan.ts
@@ -1,23 +1,17 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ScanStatus } from './scan-types';
+import { ReportData } from './report-data';
+import { ScanStatus, ScanType } from './scan-types';
 import { StorageDocument } from './storage-document';
-import { ScanType } from './website';
 
-export type WebsiteReport = {
-    reportId: string;
-    format: 'consolidated.html';
-    href: string;
-};
-
-// Summary of completed scan/crawl of full website, for one scan type (a11y, privacy, or security)
+/*
+ * Represents a scan/crawl of a full website for one scan type
+ */
 export interface WebsiteScan extends StorageDocument {
     websiteId: string; // Maps to a Website
     scanType: ScanType;
+    scanFrequency: number;
     scanStatus: ScanStatus;
-    lastScanCompletedDate?: Date;
-    report: WebsiteReport[];
-    retryCount: string;
-    // urls: string[];
+    reports: ReportData[];
 }
